app/requirements.txt
fastapi
motor
pydantic
uvicorn
python-dotenv
starlette #CORS